refactor(sessions): co-locate session-tab strip + extract Tab leaf
- Move SessionTabs → sessions/components/Tabs, rename to Tabs
- Extract inline tab JSX into Tab.tsx with full behavior contract:
  - Keyboard: Enter/Space activate, Delete/Backspace close, descendant-event suppression
  - ARIA: role=tab, aria-selected, aria-controls, aria-label with (ended) suffix
  - Agent glyph chip, accent stripe (active only), StatusDot (running/paused only)
  - Active-no-op guard for click + keyboard
- Add Tab.test.tsx (~19 tests) covering leaf scope
- Trim Tabs.test.tsx to orchestrator scope (~13 tests)
- Update WorkspaceView import path + JSX usage
- Tests: npx vitest run passing (1745 tests)
- Lint: npm run lint passing
- Type-check: npx tsc --noEmit passing
- feature_list.json: marked #7 as passing
